@@ -1,21 +1,19 @@
 source :rubygems
 
-gem 'rails',       '3.0.9'
-gem 'rake',        '~> 0.8.7', require: false
+gem 'rails',       '3.2.8'
 gem 'pg'
 gem 'sequel'
 gem 'sequel_pg',   require: 'sequel'
 gem 'yajl-ruby',   require: 'yajl'
 gem 'amatch'
-gem 'stringex',    '~> 1.3.0'
+gem 'stringex'
 gem 'redis'
 gem 'dalli'
 gem 'rack-cache',  require: 'rack/cache'
 gem 'gcoder'
 gem 'lcbo',        '1.2.3'
 gem 'rdiscount'
-gem 'haml'
-gem 'sass'
+gem 'haml-rails'
 gem 'exceptional'
 gem 'whenever',    require: false
 gem 'colored',     require: false
@@ -23,16 +21,29 @@ gem 'zippy',       require: false
 gem 'aws-s3',      require: false
 gem 'unicorn',     require: false
 
-group :development do
-  gem 'capistrano'
-  gem 'rvm'
-  gem 'awesome_print'
+group :assets do
+  gem 'jquery-rails'
+  gem 'sass-rails'
+  gem 'therubyracer'
+  gem 'uglifier'
 end
 
-group :test do
-  gem 'fabrication'
+group :development do
+  gem 'hooves',            require: 'hooves/default'
+  gem 'awesome_print'
+  gem 'annotate',          require: false, git: 'git://github.com/ctran/annotate_models.git'
+  gem 'capistrano',        require: false
+  gem 'capistrano-ext',    require: false
+  gem 'capistrano_colors', require: false
+  gem 'rvm-capistrano',    require: false
 end
 
 group :test, :development do
   gem 'rspec-rails'
+  gem 'capybara',    require: false
+  gem 'fabrication', require: false
+end
+
+group :test do
+  gem 'database_cleaner'
 end
