@@ -1,61 +1,65 @@
 require File.expand_path('../boot', __FILE__)
 
-# Select frameworks
+#require 'active_record/railtie'
 require 'action_controller/railtie'
 require 'action_mailer/railtie'
+require 'sprockets/railtie'
 
-# If you have a Gemfile, require the gems listed there, including any gems
-# you've limited to :test, :development, or :production.
-Bundler.require(:default, Rails.env) if defined?(Bundler)
+if defined?(Bundler)
+  # If you precompile assets before deploying to production, use this line
+  Bundler.require(*Rails.groups(assets: %w[development test]))
+  # If you want your assets lazily compiled in production, use this line
+  # Bundler.require(:default, :assets, Rails.env)
+end
 
 module LCBOAPI
-  def self.[](value)
-    @config ||= YAML.load_file((Rails.root + 'config' + 'lcboapi.yml').to_s)
-    @config[value]
-  end
-
   class Application < Rails::Application
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
 
-    # Middleware
-    config.middleware.use Rack::Cache,
-      :verbose     => true,
-      :metastore   => 'memcached://127.0.0.1:11211/lcboapi-meta',
-      :entitystore => 'memcached://127.0.0.1:11211/lcboapi-body'
-
-    # Using RSpec and Fabrication
+    # Generator config
     config.generators do |g|
-      g.test_framework :rspec, :fixture => true
-      g.fixture_replacement :fabrication, :dir => 'spec/fabricators'
+      g.test_framework :rspec, fixture: true
+      g.fixture_replacement :fabrication
+      g.javascripts false
+      g.stylesheets false
     end
 
     # Custom directories with classes and modules you want to be autoloadable.
     config.autoload_paths += %W[#{config.root}/lib]
 
-    # Only load the plugins named here, in the order given (default is alphabetical).
-    # :all can be used as a placeholder for all plugins not explicitly named.
-    # config.plugins = [ :exception_notification, :ssl_requirement, :all ]
-
-    # Activate observers that should always be running.
-    # config.active_record.observers = :cacher, :garbage_collector, :forum_observer
-
-    # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
-    # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
-    # config.time_zone = 'Central Time (US & Canada)'
-
-    # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
-    # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
-    # config.i18n.default_locale = :de
-
-    # JavaScript files you want as :defaults (application.js is always included).
-    # config.action_view.javascript_expansions[:defaults] = %w(jquery rails)
-
     # Configure the default encoding used in templates for Ruby 1.9.
-    config.encoding = 'utf-8'
+    config.encoding = 'UTF-8'
 
     # Configure sensitive parameters which will be filtered from the log file.
     config.filter_parameters += [:password, :password_confirmation]
+
+    # Enforce whitelist mode for mass assignment.
+    # This will create an empty whitelist of attributes available for mass-assignment for all models
+    # in your app. As such, your models will need to explicitly whitelist or blacklist accessible
+    # parameters by using an attr_accessible or attr_protected declaration.
+    # config.active_record.whitelist_attributes = true
+
+    # Enable the asset pipeline
+    config.assets.enabled = true
+
+    # Define asset manifests
+    config.assets.precompile -= %w[
+      application.css
+      application.js
+    ]
+    config.assets.precompile += %w[
+      lcboapi.css
+      lcboapi.js
+    ]
+
+    # Version of your assets, change this if you want to expire all your assets
+    config.assets.version = '1.0'
+  end
+
+  def self.[](value)
+    @config ||= YAML.load_file((Rails.root + 'config' + 'lcboapi.yml').to_s)
+    @config[value]
   end
 end
