@@ -1,3 +1,3 @@
 Mime::Type.register 'application/vnd.google-earth.kml+xml', :kml
-Mime::Type.register 'application/zip', :zip
+# Mime::Type.register 'application/zip', :zip
 Mime::Type.register 'text/tsv', :tsv
